Remove enterprise page client component
